requirements.txt
opencv-python
mediapipe
pyautogui
numpy
pystray
pillow
